+++ changelog.md
@@ -6,6 +6,7 @@
 * Chinese language update (Thanks to @binaryoung)
 * Allow custom Mailer types (thanks to @michield)
 * Cope with spaces around SMTP host specs
+* Fix processing of multiple hosts in connect string
 
 ## Version 5.2.7 (September 12th 2013)
 * Add Ukranian translation from @Krezalis